+++ apps/forms/components/utility/disable-swipe-back.tsx
@@ -1,0 +1,13 @@
+import React, { useEffect } from "react";
+
+export const DisableSwipeBack = ({ children }: React.PropsWithChildren<{}>) => {
+  useEffect(() => {
+    document.body.style.overscrollBehaviorX = "none";
+
+    return () => {
+      document.body.style.overscrollBehaviorX = "";
+    };
+  }, []);
+
+  return <>{children}</>;
+};
